check if file exists before delete it

+++ src/Model/File.php
@@ -167,14 +167,14 @@ class File extends Model {
 
 	public function deleteFromDisk(){
         $filePath = $this->getRelativePath();
-        unlink($filePath);
+        if( file_exists($filePath) ) unlink($filePath);
 
         $fileName = pathinfo($filePath, PATHINFO_FILENAME);
         $files = scandir($this->relativeFolder . $this->folderID . '/');
         foreach($files as $file){
             if( StringUtils::startsWidth($file, $fileName) ){
                 $suffixFilePath = $this->relativeFolder . $this->folderID . '/' . $file;
-                unlink($suffixFilePath);
+                if( file_exists($suffixFilePath) ) unlink($suffixFilePath);
             }
         }
 	}
